# [Enhancement] Optimize `select count(*)` query pattern for external table