--- a/be/src/formats/parquet/group_reader.cpp
+++ b/be/src/formats/parquet/group_reader.cpp
@@ -90,7 +90,7 @@ Status GroupReader::_build_chunk_on_use_any_column(ChunkPtr* chunk, size_t* row_
     ctx.row_count -= count;
     *row_count = count;
 
-    ColumnPtr c = ColumnHelper::create_const_null_column(count);
+    ColumnPtr c = ColumnHelper::create_column(ctx.slot_desc->type(), ctx.slot_desc->is_nullable(), true, count);
     (*chunk)->update_column(c, ctx.slot_desc->id());
     return Status::OK();
 }