--- a/be/src/formats/parquet/group_reader.h
+++ b/be/src/formats/parquet/group_reader.h
@@ -181,7 +181,6 @@ private:
 
     DictFilterContext _dict_filter_ctx;
 
-
     struct UseAnyColumnContext {
         size_t row_count = 0;
         SlotDescriptor* slot_desc = nullptr;
@@ -190,7 +189,7 @@ private:
     bool _can_use_any_column() const;
     Status _build_chunk_on_use_any_column(ChunkPtr* chunk, size_t* row_count);
     void _init_use_any_column();
-    UseAnyColumnContext _use_any_column_ctx;    
+    UseAnyColumnContext _use_any_column_ctx;
 };
 
 } // namespace starrocks::parquet