… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleSet.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleType.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/implementation/HiveScanImplementationRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/LabelMinMaxCountOnScanRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/PruneHDFSScanColumnRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/PruneScanColumnRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/RewriteMinMaxCountOnScanRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/plan/PlanFragmentBuilder.java
… gensrc/thrift/PlanNodes.thrift