Fixes #issue

## Summary

This pattern can be observed during pruning columns from scan operator. When there is no materialized column required from a table, which means any column could suffice(and any value could suffice).

So I add a `canUseAnyColumn` attached to scan operator and scan node.

And if a scan node has an attribute that `can_use_any_column = true`, I can just return any const column with the expected size.

Right now the optimization works for parquet only. It should be very easy to extend them to other formats and other data sources.

## Benchmark

Use tpch-1t parquet on a single node for example. With this optimization, the following query `select count(*) from lineitem` reads about 
- data: 133MB
- count: 8.5K
- takes: 5.7s

<img width="317" alt="image" src="https://github.com/StarRocks/starrocks/assets/1081215/f674f71a-1808-4fd4-93dd-88be35cf7a49">


```
MySQL [tpch_parquet]> select count(*) from lineitem;
+------------+
| count(*)   |
+------------+
| 5699989709 |
+------------+
1 row in set (5.760 sec)
```

If we use `select count(l_partkey) from lineitem`, it reads about
- data: 21GB
- count: 14K
- takes: 10.4s

<img width="367" alt="image" src="https://github.com/StarRocks/starrocks/assets/1081215/817622f5-42e0-4f5e-bdd1-f1a9870859f7">


```
MySQL [tpch_parquet]> select count(l_partkey) from lineitem;
+------------------+
| count(l_partkey) |
+------------------+
|       5699989709 |
+------------------+
1 row in set (10.422 sec)
```


## Test Cases

Since `select count(*)` is a very common pattern, there is no need to add test cases for that optimization.

## What type of PR is this:
- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4