@wanpengfei-git 2023-06-23
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 19 / 22 (86.36%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/plan/PlanFragmentBuilder.java|2|4|50.00%|[3107, 3108]|
|:large_blue_circle:|com/starrocks/planner/ScanNode.java|2|3|66.67%|[80]|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/logical/LogicalScanOperator.java|4|4|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/PruneHDFSScanColumnRule.java|6|6|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/implementation/HiveScanImplementationRule.java|1|1|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalScanOperator.java|3|3|100.00%|[]|
|:large_blue_circle:|com/starrocks/planner/HdfsScanNode.java|1|1|100.00%|[]|